+++ CHANGES.rst
@@ -9,7 +9,9 @@ v3.1.0 (xxxx-xx-xx)
 
 - Add ``fill``.
 - Add ``in_range``.
+- Add ``matches_property``.
 - Add ``start_case``.
+- Make callbacks support ``matches_property`` style as ``[key, value]``.
 
 
 .. _changelog-v3.0.0: